training models/disease detection/requirements.txt
streamlit
tensorflow
numpy
pillow
scipy